+++ templates/9_quast.tera.sh
@@ -64,9 +64,9 @@ if [ -e 1_genome/paralogs.fa ]; then
     QUAST_LABEL+="paralogs,"
 fi
 
-if [ -e 1_genome/repetitives.fa ]; then
-    QUAST_TARGET+=" 1_genome/repetitives.fa "
-    QUAST_LABEL+="repetitives,"
+if [ -e 1_genome/repetitive.fa ]; then
+    QUAST_TARGET+=" 1_genome/repetitive.fa "
+    QUAST_LABEL+="repetitive,"
 fi
 
 QUAST_LABEL=$( echo "${QUAST_LABEL}" | sed 's/,$//' )