+++ templates/9_stat_final.tera.sh
@@ -20,9 +20,9 @@ if [ -e 1_genome/paralogs.fa ]; then
     printf "| %s | %s | %s | %s |\n" \
         $(echo "Paralogs"; faops n50 -H -S -C 1_genome/paralogs.fa;) >> statFinal.md
 fi
-if [ -e 1_genome/repetitives.fa ]; then
+if [ -e 1_genome/repetitive.fa ]; then
     printf "| %s | %s | %s | %s |\n" \
-        $(echo "Repetitives"; faops n50 -H -S -C 1_genome/repetitives.fa;) >> statFinal.md
+        $(echo "repetitive"; faops n50 -H -S -C 1_genome/repetitive.fa;) >> statFinal.md
 fi
 
 # anchors