+++ CHANGELOG.md
@@ -80,7 +80,7 @@
     * --gluemin 10
     * --fillmax 500
 * Support SAMN and PRJNA in `anchr ena`
-* Add `repetitives.fa` to `1_genome`
+* Add `repetitive.fa` to `1_genome`
 * Add docs
     * fda_argos.md
     * yeast.md